Add files via upload
Updated: 1/26/25
My scratch project consists of two balloons that are popped with a dart. The dart is a draggable sprite that interacts with the two balloons. When the dart is dragged over either balloon a "popping" sound is omitted and the balloon is randomly repositioned. Each ballon is independent with three different costumes.